Sistemas de Cifrado, Últimos SignIn y reparación de Event
Avance en el sistema de cifrado y gestión de correos.

Añadida entidad para controlar los últimos SignIn

Por error en el ultimo commit-push de esta rama se añadieron cambios temporales de ficheros de otra rama que se han deshecho.
