+++ src/java/cypher/Mail.properties
@@ -1,7 +1,9 @@
-SUBJECT = Recuperaci\u00f3n de contrase\u00f1a solicitada
+SUBJECTRESET = Recuperaci\u00f3n de contrase\u00f1a solicitada
 
-BODY1 = Esta es su nueva contrase\u00f1a
+BODY1 = <p>Esta es su nueva contrase\u00f1a</p></br>
 
-BODY2 = Si usted no ha solicitado un cambio de contrase\u00f1a, contacte con soporte.
+BODY2 = </br><p>Si usted no ha solicitado un cambio de contrase\u00f1a, contacte con soporte.</p></br></br><p>Un Saludo,</br>Administraci\u00f3n del sistema Rental2G1C</p>
 
-CAMBIOPASSWORD = Ha realizado un cambio de contrase\u00f1a de forma exitosa. Si no ha sido usted, contacte soporte lo antes posible.
+SUBJECTCHANGE = Recuperaci\u00f3n de contrase\u00f1a solicitada
+
+CAMBIOPASSWORD = <p>Ha realizado un cambio de contrase\u00f1a de forma exitosa.</br>Si no ha sido usted, contacte con soporte lo antes posible.</p></br></br><p>Un Saludo,</br>Administraci\u00f3n del sistema Rental2G1C</p>
